requirements.txt
python-telegram-bot==13.7
RPi.GPIO==0.7.0
youtube_dl
pyyaml
fsnotify
